refactor CryptoBase
